Add QueryPage-based poor man's queue API
Add Special:ImageLabeling query page and corresponding
action=query&list=querypage&qppage=ImageLabeling for
listing pages with pending labels.

QueryPage is not great for this: it is meant for
more-or-less static reports such as backlog lists, not
microcontribution interfaces. It's good enough to get us
off the ground, though, and saves a day or two of
development time for now.

Images are sorted by sha1 (ie. essentially randomly), which
is not nice but it's the only way to select images with
pending labels in a performant way. (Sample explain: P8810)
In theory sorting on mvl_timestamp would be another option
but all my attempts to do that resulted in a filesort.

No patchfile is provided for the schema change since the
extension is still in development. The command to update is:
    ALTER TABLE machine_vision_label ADD COLUMN mvl_review tinyint NOT NULL DEFAULT 0 AFTER mvl_wikidata_id;
(plus indexes but those should not matter on a developer setup)

Change-Id: I57195e70d61f60713e00e64205698e64b689bffb
Depends-On: I8b67268e3c59a51ed00bd832f80df184dd771c3b