Fix: Pass in database when constructing NameTableStore
Previously no entry was created in the machine_vision_provider table for
a provider for which an ID was acquired with NameTableStore. Apparently
the association was persisted somewhere, though, which created problems
when debugging. NameTableStore is not well documented
(https://phabricator.wikimedia.org/T216468) so I'm not sure what was
going on there.

With this update we pass in the database we want, which
results in the entry being persisted in machine_vision_provider as
expected.

Change-Id: I5fc3d7707271e75de7c6f9029b6587242b645ccc