Resolve Wikidata item labels over HTTP for development
Wikidata item labels should be resolved in PHP. However, for genuine
Wikidata IDs, we can't resolve them internally via EntityLookup outside
of WMF production, since we won't have Wikidata set up as our Wikibase
repo.

This adds follow-up wbgetentities requests to the Wikidata public API
for item labels. This functionality is to be used only for testing; in
production, we should use EntityLookup as implemented in
https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/MachineVision/+/530594/

Change-Id: I784e7013ac2bab5d3b0629dc838275c3bc76ea39