Allow filtering suggestions for review by uploader
Adds a new API list module, ApiQueryUnreviewedImageLabels, that returns
titles of images with labels needing review. This supersedes
SpecialImageLabeling, which will be removed in a patch to follow.

The query for results will be much the same as before, but with
specifying 'generator=unreviewedimagelabels' rather than 'generator=
querypage&gqppage=ImageLabeling'.

To filter by uploader, specify a user ID in the 'uploader parameter'.

Limits results to between 1 and 10 per query, with a default of 1. This
is based on the idea that we are moving away from a "feed" and toward
one-at-a-time voting.

Bug: T231104
Change-Id: I131729f8d5e439868b244a5b55f01ce79054d5df