Dev: Register WikibaseClient as an extension dependency
Adding just Wikibase as an extension dependency doesn't work, because
there is no Wikibase extension; the extensions are WikibaseClient and
WikibaseRepository.

This adds WikibaseClient as a dependency. Hopefully, it will allow us to
remove the Gerrit configuration that results in all of the Wikibase
ecosystem's sometimes slow and flaky test suites running for each of our
changes.

Depends-On: If13d2ee98f1f748d7b1553860ef22c82b4b06100
Change-Id: I9d16b01995f5999a0fb293df8cbaa1941c52ce3b