Add headless JS unit tests using mw-node-qunit library
We've been experimenting with headless node-qunit tests in the MediaInfo
extension, and ended up building a custom testing framework to support our
semi-unorthodox way of using OOUI (templates and render methods, require()
statements with ResourceLoader, etc).

A lot of what we ended up doing there is pretty similar to what the web team has
done with their mw-node-qunit library. Now that we are developing a new
extension, I wanted to see if I could lean on this library for our testing needs
in the hopes of eventually standardizing on one way of running headless JS unit
tests.

Change-Id: Ifa8c3e3a3a30f38775ca786f96c43826f2d07d8b