Refactor pager to follow cardstack design
The pager element, which wraps all content within this page at the
moment, previously showed a set number of images at a time with a
"more" button to show the next batch. To match the latest designs,
this commit refactors that element to present a single image at a
time and allow paging through images individually. The "page"
element, which housed a single batch of items, has been removed.

Bug: T231926
Change-Id: If7985064fcadc3627a28209fb1a936c74689c00c