Add script to create file lists based on global image usage
Adds a script to create file lists based on global image usage. The
minimum number of external usages and the namespaces to be considered
are passed in as parameters (with NS_MAIN as the default for the
latter).

Filtering is applied as with the script to create lists based on
categories and templates.

Bug: T234250
Change-Id: I76d5757abf7e2052eddd34791bc12b021930d201