Refactor suggestions and suggestions group
This patch removes the SuggestionsGroupWidget component completely in
favor of storing a SuggestionWidget's confirmed status in the widget
itself rather than in arrays in the ImageWithSuggestionsWidget. This
change has the added benefit of enabling better keyboard navigation
since suggestions don't lose focus when they're toggled as we're no
longer re-rendering all suggestions when one changes.

Bug: T234964
Change-Id: I67947860dc7b784a899fc68a4fe0cc1d30763f0a