Provide for withholding "NSFW" images from being reviewed
Provides for configuring likelihood limits along the SafeSearch
dimensions such that, if an image scores greater than or equal to the
configured likelihood level on that dimension, labels are stored with an
initial state of REVIEW_WITHHELD and are not shown for review.

N.B. This requires an update to the Google Cloud Vision handler config,
if used, to pass in an additional parameter to reflect desired
SafeSearch limits. The README file is updated accordingly.

Bug: T234250
Change-Id: I8427d6ba4083f533f5df515d4da35a04e47cbd05