Allow URL hash to determine active tab
Updates the SuggestedTagsPage class to keep the URL hash in sync with the
currently active tab. Changing between tabs will update the URL hash
accordingly, switching between #user and #popular. The SuggestedTagsPage will
also listen for URL hash changes from the user and will switch tabs accordingly
if the new hash corresponds to a tab that exists in the layout. Finally, an
optional startTab config param can be provided which determines what tab should
initially be shown to the user; default is "popular".

In all cases of hash manipulation, history.replaceState() is used so as to not
clutter the users history with redundant entries.

Bug: T235214
Change-Id: I90d50abaec83fca25afd9d5e1ac87497c08cb4ff