Move query to top-level component
To avoid collisions between popular and user queries and to ensure
the user is always seeing an up-to-date collection of images, this
patch moves the API call for unreviewed images to the top-level
component, which passes results down to the child cardstack components.

Any time the user changes tabs, a new query will run.

Bug: T235703
Change-Id: Ief57f83080bd9b81b82d44d9ef23fb79f7c33b8c