Link image labels to filepage
The image labels API includes a "descriptionurl" property in the imageinfo
object that is returned for each image under review.

This change adds "descriptionurl" to the ImageData model class and uses it as a
label link in the ImageWithSuggestions widget. The actual image element is also
wrapped with a link tag which leads to the same URL.

Bug: T234962
Change-Id: Ife31e3d103801625a72236016f9c85ee110e5483