Avoid generic 'moduleID' global variable
The generic nature of this name makes it hard to interoperate with
other page-agonistic code that may load on the same web page.

The way we generally use mw.template in MediaWiki is to specify
the module name directly. Abstracting this is an interesting idea
and would perhaps be done by injecting it as a "data.json" key,
which this patch could be adapted to, if preferred.

Bug: T230813
Change-Id: I0318f62a2b6a6899343d26dd6d6f22e908d50e55