Add throttling support for the label fetcher script
It's necessary to provide for throttling requests to third-party
APIs in order to respect any request quotas that exist. For example, the
Google Cloud Vision API has a quota of 1,800 requests per minute. This
patch adds support for configuring a maximum number of requests per
minute in the handler.

A default of 0 (no throttling) is provided, so there's no need for local
configs to be updated except for testing this.

Bug: T236311
Change-Id: I84142a2757e16ab3962dcfd7bc6ac4a1d12c2726