Optionally limit Special page access to testers only
Adds a new 'imagelabel-test' user right, and sets that permission on
Special:SuggestedTags if the extension is configured to enforce it. This
right will apply to members of the new 'machinevision-tester' group.

The intent is for this permission to be granted to specific users during
the user testing period prior to official launch.

Bug: T236416
Change-Id: I62a63bd7fe5053da6f11db8f12de5e5c1f05ba38