Condiitonally withhold remaining unreviewed titles on save complete
Adds a hook handler for PageContentSaveComplete that checks the edit
summary for the characteristics of a depicts statement addition (namely
the substrings 'wbsetclaim-create' and the configured depicts property
ID string). If this check passes and the title is in the File namespace,
a deferred update is added to recheck the title with TitleFilter and set
any remaining unreviewed labels to REVIEW_WITHHELD if isGoodTitle fails.

Bug: T236408
Change-Id: I2af05c4b76dec4aa65e7fec158751f80cf0c3775