Use internal HTTP request abstractions to make Google requests
Using Google's Cloud Vision client library was imposing certain
limitations. Most importantly, it has no mechanism for specifying an
HTTP proxy, which is a requirement in production. Also, it expects
the credentials file location to be specified as an environment
variable, which is not ideal. Finally, it is beta software.

This commit swaps out the client library in favor of constructing and
making requests with MediaWiki's own MwHttpRequest mechanism. This
affords us more flexibility to interact with the Cloud Vision API in
ways that suit our production environment.

The google/auth package is still used for making auth token requests.

NOTE: This patch requires configuration changes as reflected in the
changes to the README file. Specifically, the services required to be
passed into the GoogleCloudVisionHandler have changed, and the path to
the Google credentials file is now specified as an extension
configuration option rather than an environment variable.

Bug: T236843
Change-Id: I155bf11b85ae29c13cceb5076133260212a284ab