Performance: Execute labeling requests in a Job, not a DeferredUpdate
As discussed on I66dec087c282554edf04d5eab0bad67bcb0a8e38, the image
annotation request and storage operation is fairly heavy. It is also not
time-sensitive. Accordingly, this patch updates it to happen in a JobQueue Job.

Note: This patch introduces substantial changes to the extension
configuration and testing workflow! Please review the README updates for
details.

Bug: T230813
Change-Id: I8d8bc82fa36cc3dc2429687183bf22ccdd39c850