Send a thumbnail URL for labeling if the original file size is too large
Google Cloud Vision imposes a max image size of ~10MB. If the image is
larger than the max image size, send a thumbnail that fits in a
1280x1024 bounding box. This is a size that is among the thumbnail
buckets created on upload, and will avoid undue burden on the
thumbnailers.

TitleFilter is updated to no longer filter images based on their size.

Change-Id: Icffcbdfb912036a2ea17f9446851a1a98155cf63