Fix: Do not return after inserting a single suggestion
A misplaced return statement recently introduced into
Repository::insertLabels() is causing only a single suggested label to
be inserted into the DB. This puts the return statement in the correct
place.

Also added a check for the expected keys in the response object, in an
abundance of caution.

Change-Id: I8a3a9ee9373e4cb92fcf6daf5d2133680bb0e030
Hygiene: Remove a stray space.