Handle case where entity not found for given ID
In production, an error is occuring where the EntityResolver is not able
to find an entity with the provided ID. This patch handles this case and
logs a warning for further investigation into the underlying cause.

Hygiene: Update a couple of functions to be more procedural and less
functional for easier reading.

Bug: T238245
Change-Id: I2ad9007118cc34c742d12040296c14dc88fa7468