Show an appropriate message when the user has no uploads
Currently the messages shown when a user has no uploaded images with
tags ready for review assume that the user has already labeled images.
A different design was created for when a user has not yet uploaded any
images with labels available for labeling, but it has not yet been
added. This patch adds it.

Change-Id: I8599b5dec6e6bc90a2217488a85f21bd5107883d