Stop passing in a base rev when creating a MediawikiEditEntity
MediawikiEditEntity::newEditEntity takes an optional $baseRevId
parameter that allows specifying a base revision ID for checking for
possible edit conflicts. That checking is not needed here because the
only editing possible through the feature—the creation of an SDC Depicts
statement—is strictly additive, and so the only conflict possible is for
the Depicts statement in question already to have been added, which is a
case that's already handled.

Bug: T238567
Change-Id: I2eec6860af91a60859fb7d4c67840d75bf7c5390