Make tag review validation more permissive
Currently we fail all of a user's votes and die with an error if the
user has submitted a vote for a label that's already been reviewed. This
is unnecessarily strict and likely to cause trouble given the removal of
suggestion randomization.

Instead, we'll log a warning and conditionally allow the vote to go
through. Either the vote will be the same as the previous voter's vote,
and nothing will change; the later voter will approve a previously
rejected tag, and a Depicts statement will be added as a result of the
later vote, which is acceptable and arguably an improvement; or the later
voter is attempting to reject a previously accepted tag, in which case
we'll prevent the vote from being processed any further.

Change-Id: Ice62280a12a0c617008e3ce7dc2abb8cab28cdbd