Add no-JS fallback message to Special:SuggestedTags
Some message should be shown to a user with JavaScript disabled or
otherwise unsupported upon loading Special:SuggestedTags; currently, we
show a blank page. This adds an error box displaying a simple message
that JavaScript is required in order to use the Computer-Aided Tagging
feature.

The message is included by default, and hidden by a CSS snippet
included in the BeforePageDisplay hook. The BeforePageDisplay hook is
used to prevent a FOUC where the message would otherwise appear in all
cases before content is fully loaded.

This initial snippet is only included for pages that know about the
ext.MachineVision module (i.e., only Special:SuggestedTags) to avoid
loading unnecessary styles on unrelated pages.

Bug: T240210
Change-Id: Id480264494fbe8c6fab8db018b0bccbfd0fcee98