Don't show focus outline on suggestions unless user presses tab
To avoid showing mouse users the browser-provided focus outline,
add a style to suggestion widgets to remove the outline, then
listen on the parent widget for tab key usage. On tab, a class on
the parent widget will be removed and the style will no longer
apply to suggestion widgets, showing the outline to tab users.

Also, prefer event.key and keydown over event.keyCode and keypress

Bug: T240510
Change-Id: I6b278d22f4429ab8cba2d9f3b18c1a820a82e2be