Fix: Bail out of label fetching job if local file not found
A configurable delay is applied to annotation fetching jobs due to the
high likelihood of newly uploaded files being deleted. However, the job
currently throws an error due to the assumption that the file will be
found locally. This updates the job to log the file title and bail out
in the event that the file is not found at job execution time.

Bug: T240733
Change-Id: Ic861280e889f128fd108f35a165d4f3be2ca90ae