Update random unreviewed image query to directly return image names
Adds a join on the image table to remove the need for a separate query.

NOTE: This tweak is introduced as a separate change from the patch
removing DISTINCT since it's not required as part of the essential query
performance fix, and not guaranteed to have a positive impact. I've kept
it separate so that we can assess its impact independently from the
fundamental, required fix.

Change-Id: Ib7fb887a5d80701d0558d105ff634a04f0a8b295