Assign images to original uploader
Currently images show up on the personal uploads tab of the user
who made the last revision to the file. This is confusing in the
case of admins making reverts, a common occurrence, then seeing
those images on their personal uploads tab. This patch looks for
previous versions of a file and, if found, grabs the user who
uploaded the oldest one.

Bug: T245212
Change-Id: I0657c0f91670d996cc5ed8d845ce48815956a1c6