Add descriptions and aliases to label data from API
Updates the LabelResolver class to fetch the description and first alias item
for each label suggestion. This is done for both code paths (internal and
external wikibase items). Also updates the initial data passed in to JSConfigVars
to include these values and makes a small change in the info action handler to
accommodate the new data structure that LabelResolver->resolve() returns.

The shape of an "imageLabel" item returned by the MV API now looks like this:

	"wikidata_id": "Q1255864",
	"state": "unreviewed",
	"label": "fun",
	"description": "enjoyment of pleasure",
	"alias": "😜"

Note: this change does *not* actually touch any front-end code, so the new data
is not yet displayed to the user. This will be handled in a follow-up patch.

Bug: T249133
Change-Id: I1e53b787a45c14d1fe437ebd6d5dabed7a9e04a5