CAT: Return only "assessed" images if no user ID is provided
Updates Repository::getTitlesWithUnreviewedLabels to return only titles
that had labels added as part of the "assessed" (featured/valued/quality)
image group if no uploader ID is passed in.

Since we don't currently have any direct way of identifying "assessed"
images/labels at present, we have to resort to a hack: we'll select those
with label suggestions entered into the database prior to December 1, 2019.
This is after labels were requested for the "assessed" group, but before
label requests were enabled for new uploads.

Ultimately (sooner rather than later), the DB schema should be updated to
include group membership. This will be needed to support future labeling
groups, such as those selected based on on-wiki usage. This temporary fix
is to improve the quality of the feature in the meantime.

No tests are added for this change because MediaWiki provides no
affordances for testing upload behavior (and associated changes to the
image table) with reasonable effort. However, it works correctly based
on manual testing.

Bug: T249270
Change-Id: I2fbc823a8ac458ec71925d3eed5f479a0ae0aedb