Self-bundle notifications
Even though the current notifications have been configured to bundle,
Echo's built-in bundling unbundles once notifications have been read,
which could easily cause some polliution in the notifications list,
given a sufficient amount of uploads.

Instead of relying on the built-in bundling, we'll simply limit the
amount of notifications: notifications will not be delivered to users
when they already have one of this kind and it's still unread.

Bug: T248322
Change-Id: I63d1de289bcf499c2b4468bdd2834cd5a736d581