Add support for WITHHOLD_ALL review state
Adds a new WITHHOLD_ALL review state, represented in the database by the
integer -3. Images will have all of their suggested labels stored with
this state when they come back with one or more suggested labels that
appear on a list of terms indicating an image that should be withheld
from both the popular and personal uploads views in
Special:SuggestedTags.

The intended use case for this update is to withhold all images
containing humans from both views.

Bug: T249939
Change-Id: I27aca4ab58e38f481f6e5754f7f6002cdebf7c9f