Fix: Apply blocklist filtering after determining withholding state
Currently, a Wikidata ID should not be on both the list of tags for
removing and the list of tags that indicate an image should be withheld
from Special:SuggestedTags, because the tag filtering is applied first.
An image could mistakenly be retained for presentation if a tag meant to
indicate it should be withheld is filtered out first. However, a
Wikidata ID should be able to appear on both lists.

This patch reverses the order, so that the initial review state (which
indicates the withholding state) is determined prior to any tag
filtering. A new check is added after the filtering step to bail out if
the list of suggested labels is empty after filtering.

A side benefit of this is that blocklist filtering is moved out of the
Repository class, where it was applied somewhat behind the scenes when
fetching mapped Wikidata IDs, and into the GoogleCloudVisionClient where
it is applied along with the rest of the GCV API response processing code.

Change-Id: I82ea30e10a6806151f88bceef94a7dce95cae9bf