Fix Repository bug potentially omitting labels from API results
A recent change to Repository::getLabels resulted in the possible
overwriting of a label suggestion if the same suggestion is returned
from the database for more than one image. Due to this bug, a label
suggestion can at present only be returned once per batch of filenames
submitted to the API.

As a result of not being presented, such labels are not voted upon,
leading to possible data integrity issues given the current assumption
that the suggested labels for an image will either be voted-upon or not
as a group.

This issue was caused by using the Wikidata ID (Q-number) alone as an
array key used for potentially consolidating multiple rows (to provide
for multiple confidence scores from different providers for the same
suggestion). As updated, the image sha1 and label Wikidata ID are
combined to create a key.

Change-Id: I9c7c27427ca302dd4cedea04b37282cf27712744