Unreviewed count should also decrement when setclaims failed
There could still be data lost when wbsetclaim fails, but
at least the UI will accurately reflect the number of remaining
images.

Alternatively, we could do both API requests async
(reviewimagelabels after setclaim succeeded) but the odds of the
call failing are low enough, the severity also low enough (it's
not like we lose a massive amount of work or so) and the cost
high enough (it'd slow submission of every image), that I think
we ought to continue to send both API requests async.

Bug: T253749
Change-Id: I302c36c7094c14a21f71a6994264714d439a922f