Fix scroll on skip
Previously, when the user skipped an image, the container would
momentarily collapse, sending the user to the top of the page so
that they'd have to scroll back down to the image card. Instead,
we'll add the new image card before the old one is removed and
hide it from view so that there's never an instant during which
there is no image card component present.

Change-Id: I18edd713466cfc59c288a05274c4ad4b8d84317b