Add expanded tags UI with suggestion aliases and descriptions
This patch adds a toggle switch base component, which is used in
the Suggested Tags interface to allow the user to toggle between
the normal tag state and expanded tags, which show the relevant
wikidata item's first alias and description.

As part of this patch the categories list and the suggestions
group were abstracted out into their own components to keep
the ImageCard component from growing too large.

Bug: T245023
Change-Id: I290cface597fc10b2f5d71ea6fdbd731986c7bc0