Start using priority column when selecting images to review
This will start to use the new priority column to figure out
which images to return first (while still keeping some of the
randomization)

The priority column values are currently not being set, so
they'll still default to 0. There will be a separate patch
to feed additional data with different priorities.

Bug: T250748
Change-Id: I23fdda574bb16d9d51ec9ccd6c23b01371c8aa3f