eslint: Upgrade eslint-config-wikimedia, use mediawiki ruleset
Use the wikimedia/mediawiki eslint ruleset (which includes Vue rules),
and remove a bunch of rules that that makes redundant.

Remove names from component definitions: we don't use them (they're only
for global registration, which we don't do) and eslint was complaining
about 'Button' being an invalid component name (because <button> is an
HTML tag).

Add disable comments for some other rule violations. Most of them are
false positives and should be disabled, except for the dynamic message
in App.vue.

Change-Id: I4530a28cd0819207fd0a251bb42826759d4e3fe0