Remove unused method
- it is not currently assigned any value anywhere
- it is not currently called anywhere

I assume this was superseeded by isTooManyRequestsError
and backoffSeconds, and we simply don't need this anymore;
it certainly doesn't look trivial to incorporate into the
current structure of fetchSuggestions, where multiple
handlers (theoretically, not that we have them) handlers
could have different limits.

In order to prevent confusion down the road
("I set/changed this value but it's not doing anything"),
I'm just going to remove this method.

Change-Id: Iddeaa9d85e28a2c87480ce7b721480d1dc59a693