Stop using prophecy in tests
We want to stop using this library, compare T315892 and T315966. (I
don’t think it’s worth creating a separate MachineVision task.)

This is a complete shot in the dark, totally untested, because all the
tests in this test class are currently skipped, and have been for two
years. So this is mainly to get the prophesize() call out of codesearch.

Change-Id: Iff2f131f7caeb0ef3ab435b42be054dbf1d766f5